backend/pyproject.toml
[project]
name = "gapmeasure-backend"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.112.2",
  "uvicorn[standard]==0.30.6",
  "python-multipart==0.0.9",
  "numpy==2.0.1",
  "opencv-contrib-python-headless==4.10.0.84",
  "pydantic==2.8.2",
  "pydantic-settings==2.4.0",
  "Pillow==10.4.0"
]

[tool.uvicorn]
factory = false
